+++ src/ATAP.Utilities.BuildTooling.ProGet.PowerShell/ATAP.Utilities.BuildTooling.ProGet.PowerShell.psm1
@@ -46,6 +46,7 @@ Export-ModuleMember -Function @(
   'Resolve-PromotionTierFromFeedName'
   'Set-FloatingPackagePins'
   'Set-PSModuleFileSignature'
+  'Sync-ProGetPowerShellModules'
   'Test-PSModulePackageSignature'
   'Test-ProGetFeedSet'
   'Test-PromotionWithinCeiling'